Move handler test helpers out of the handlers package
