+++ handlers/handlertest/handlertest.go
@@ -1,4 +1,4 @@
-package handlers
+package handlertest
 
 import (
 	"bytes"
